requirements.txt
torch
transformers
python-docx
reportlab
streamlit